# attendance_tracker_frontend/student-attendance-tracker-216606-216615/backend/node-api/openapi.yaml
openapi: 3.0.3
info:
  title: Attendance API
  version: "1.0.0"
  description: Node.js + Express + TypeScript API for attendance tracking with JWT auth and Postgres.
servers:
  - url: http://localhost:8080
x-docs:
  helper: "/api/docs"
  sse_note: "Use GET /api/sse/attendance with Bearer token"
tags:
  - name: Auth
    description: Authentication endpoints
  - name: Attendance
    description: Attendance operations
  - name: Realtime
    description: Server-Sent Events streams
paths:
  /:
    get:
      summary: Health
      tags: [Auth]
      responses:
        "200":
          description: API health and metadata
  /api/docs:
    get:
      summary: Docs helper
      tags: [Auth]
      responses:
        "200":
          description: OpenAPI info and SSE usage note
  /api/auth/login:
    post:
      summary: Login
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 4 }
              required: [email, password]
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  user:
                    type: object
                    properties:
                      id: { type: string, format: uuid }
                      name: { type: string }
                      email: { type: string, format: email }
        "401":
          description: Invalid credentials
  /api/attendance/mark:
    post:
      summary: Mark attendance
      tags: [Attendance]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                studentId: { type: string, format: uuid }
                date: { type: string, pattern: "^(\\d{4}-\\d{2}-\\d{2})$" }
                period: { type: string }
                status: { type: string, enum: [present, absent] }
              required: [studentId, date, period, status]
      responses:
        "201":
          description: Created
        "400":
          description: Validation error
        "401":
          description: Unauthorized
  /api/attendance/daily:
    get:
      summary: Daily records
      tags: [Attendance]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: date
          required: true
          schema:
            type: string
            pattern: "^(\\d{4}-\\d{2}-\\d{2})$"
      responses:
        "200":
          description: List of records
        "400":
          description: Validation error (missing or invalid date)
        "401":
          description: Unauthorized
  /api/attendance/weekly:
    get:
      summary: Weekly records
      tags: [Attendance]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: weekStart
          required: true
          schema:
            type: string
            pattern: "^(\\d{4}-\\d{2}-\\d{2})$"
      responses:
        "200":
          description: List of records
        "400":
          description: Validation error (missing or invalid weekStart)
        "401":
          description: Unauthorized
  /api/sse/attendance:
    get:
      summary: Realtime attendance SSE
      tags: [Realtime]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: text/event-stream with attendance events
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT